Create Uninstall Adobe Acrobat.sh
This will try and uninstall Adobe Acrobat for mac. if the user deleted the inbuilt uninstaller this will put the uninstall back (only past installation media i could find to capture the uninstallers were Acrobat DC, X, 2017, 202) which is required for the uninstall.  Also if Acrobat is detected in the logged in users home directory this is removed as well.